suma/pyproject.toml
[tool.poetry]
name = "suma"
version = "0.1.0"
description = ""
authors = ["gcca <git@gcca.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
alembic = {extras = ["tz"], version = "^1.13.1"}
pre-commit = "^3.7.1"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
